refactor: move TimeoutSeconds from number to time module
